Refactor: 詳細画面のズーム・パンロジックをViewからViewModelに移動
DetailViewに直接書かれていたピンチズーム・ドラッグ・ダブルタップの
状態管理とロジックをDetailViewModelに移動し、MVVM設計に準拠させた。
未使用のpos struct、isTapped、didTapImage、getShotDateを削除。
